Move RazorPreCompileModule.cs to Razor.Precompilation
