+++ test/WebSites/PrecompilationWebSite/compiler/preprocess/RazorPreCompilation.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.AspNet.Mvc;
+using Microsoft.AspNet.Mvc.Razor.Precompilation;
 
 namespace PrecompilationWebSite
 {
